apps/dsfplbot/requirements.txt
python-telegram-bot[job-queue]==21.3
aiohttp>=3.9.0
aiosqlite>=0.19.0
pytz>=2024.1
tenacity>=8.2.0
numpy>=1.24.0
httpx>=0.25.0
